implemented server client symetric encription

+++ header/send.h
@@ -6,6 +6,7 @@
 #include<unistd.h>
 #include<string.h>
 #include<stdlib.h>
+#include <openssl/rand.h>
 
 #ifndef MESSAGE_H
     #define MESSAGE_H
@@ -17,4 +18,4 @@
 
 int serialize_message(void* buffer, struct message *aux);
 
-void send_message(struct message *mex, struct sockaddr_in * dest_addr,int socket, int encrypt);
+void send_message(struct message *mex, struct sockaddr_in * dest_addr,int socket, uint8_t encrypt);
